requirements.txt
google-generativeai==0.5.4
pillow==10.3.0
openai==1.30.5
